+++ DataModeling/C#/Model/Attractions/AttractionTicket.cs
@@ -14,9 +14,9 @@ namespace DataModeling.Model
 
         public DateTime TicketDate { get; }
 
-        public float Price { get; }
+        public double Price { get; }
 
-        public AttractionTicket(int reservationID, int attractionID, DateTime ticketDate, float price)
+        public AttractionTicket(int reservationID, int attractionID, DateTime ticketDate, double price)
         {
             ReservationID = reservationID;
             AttractionID = attractionID;
